requirements.txt
ffmpeg-python
pydub
SpeechRecognition
pocketsphinx
transformers
nltk
spacy
gensim
sumy
ffmpeg-python
pytube
SpeechRecognition
pocketsphinx
gensim